% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{add_edges}
\alias{add_edges}
\title{Add edges from a protein-protein interaction matrix into a sampled subnetwork}
\usage{
add_edges(universe_matrix, matched_empty_matrix, nodelist, edge_indeces,
  n_edge_indeces, mode)
}
\arguments{
\item{universe_matrix}{a sparse matrix with all possible edges}

\item{matched_empty_matrix}{a sparse matric that is a subset of universe_matrix, to which edges are to be added}

\item{nodelist}{a list of possible nodes}

\item{edge_indeces}{an index of all entries in universe_matrix corresponding to an edge, in row,column format}

\item{n_edge_indeces}{number of rows in edge_indeces}

\item{mode}{either 'node' or 'edge', whether to sample one itneraction, or all interactions involving one protein}
}
\value{
a modified version of matched_empty_matrix
}
\description{
Add edges from a protein-protein interaction matrix into a sampled subnetwork
}

