% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_analysis_utilities.R
\name{bpc_notation_to_yeast_orf_pair}
\alias{bpc_notation_to_yeast_orf_pair}
\title{Converts a list of BPC notations to pairs of yeast ORFS}
\usage{
bpc_notation_to_yeast_orf_pair(bpc_list)
}
\arguments{
\item{bpc_list}{list of BPC ids}
}
\value{
two column table of Yeast ORFs corresponding to BPC Ids
}
\description{
given a list of BPC IDs such as 'MET3_F1::MET3_F3', convert to a pair of yeast ORFs
}

