% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{collapse_multigenes}
\alias{collapse_multigenes}
\title{Collapse multiple measurements for PCA file - returns mean of numeric columns in duplicate rows of a file,
with Gene.1 and Gene.2 columns as keys}
\usage{
collapse_multigenes(pca_file)
}
\arguments{
\item{pca_file}{a pca file, standardly formatted - object will be grouped by identity in the Gene.1 and Gene.2 columns}
}
\value{
a pca file with multiple measurements per ORF pair averaged
}
\description{
Collapse multiple measurements for PCA file - returns mean of numeric columns in duplicate rows of a file,
with Gene.1 and Gene.2 columns as keys
}

