% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{connectivity_graph}
\alias{connectivity_graph}
\title{Given an object from network_simulation_significance, makes a heatmap-like table representing the significance matrix}
\usage{
connectivity_graph(my_matr, my_color_list, border_colour = "grey30",
  border_size = 0.8, legend_labels = c(`-1` = "Decreased component size",
  `0` = "Expected component size", `1` = "Increased component size"))
}
\arguments{
\item{my_matr}{a matrix from network_simulation_significance, where the rows correspond to directions and the columns are conditions,
values are either 1 (significant, real value is greater than simulations), 0 (nonsigificant), -1
(significant, real value is less than simulations)}

\item{my_color_list}{a colour list used to construct the legend - a gradient will be made and the first colour
will be used for -1, the middle will be 0, and the last colour will be 1}

\item{border_colour}{the border colour to draw around each grid in the heatmap}

\item{border_size}{the width of the border}

\item{legend_labels}{vector of form c('-1'=label1,'0'=label2,'1'=label3)}
}
\value{
a plot
}
\description{
Given an object from network_simulation_significance, makes a heatmap-like table representing the significance matrix
}

