% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_analysis_utilities.R
\name{download_unzip_genome}
\alias{download_unzip_genome}
\title{Download a gzipped file and unzip it (used for genome downloading)}
\usage{
download_unzip_genome(genome_http_dir, genome_dest_dir)
}
\arguments{
\item{genome_http_dir}{Web address of genome}

\item{genome_dest_dir}{Where you want it downloaded}
}
\description{
Download a gzipped file and unzip it (used for genome downloading)
}

