% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{empirical_two_tailed_p}
\alias{empirical_two_tailed_p}
\title{Give two-tailed empirical p value from a bootstrap}
\usage{
empirical_two_tailed_p(observed_value, iterated_value_list)
}
\arguments{
\item{observed_value}{value observed in the experiment}

\item{iterated_value_list}{values obtained by re-sampling}
}
\value{
proportion of values as extreme as one observed or more
}
\description{
Give two-tailed empirical p value from a bootstrap
}

