% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment_utilities.R
\name{funcassociate}
\alias{funcassociate}
\title{Submit a query to funcassociate and get the results}
\usage{
funcassociate(query, universe, associations_file = system.file("go_map.txt",
  package = "bcPcaAnalysis"), order_mode = "ordered",
  go_type = "default_funcassociate", nametype = "ORF",
  network_mode = "nodewise", p_value_cutoff = 0.05, reps = 1000)
}
\arguments{
\item{query}{a vector of gene identifiers, or if mode is 'edgewise', a two column matrix
of gene names}

\item{universe}{a vector of all possible gene identifiers in the experiment, in the
same format as query.  If 'edgewise', PPIs must have the same order as in the query}

\item{associations_file}{the file used to build the GO associations, see examples}

\item{order_mode}{either ordered or unordered}

\item{go_type}{either default_funcassociate or slim, each has their own
ways to parse the file, see examples for the appropriate format}

\item{nametype}{either 'common' or 'ORF'.  'common' is not compatible with
the default_funcassociate mapping mode}

\item{network_mode}{either 'nodewise' or 'edgewise', 'nodewise' is the classic
GO analysis, in 'edgewise'}

\item{p_value_cutoff}{p value cutoff for the enrichment analysis, passed directly to the server}
}
\value{
The funcassociate server output parsed into a list
}
\description{
Submit a query to funcassociate and get the results
}

