% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{get_density_from_edgelist}
\alias{get_density_from_edgelist}
\title{Returns the size of the largest component from a list of edges}
\usage{
get_density_from_edgelist(edgelist)
}
\arguments{
\item{edgelist}{a two-column matrix}
}
\value{
graph density  (numeric)
}
\description{
Returns the size of the largest component from a list of edges
}

