% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_analysis_utilities.R
\name{get_genes_matching_regex_motif}
\alias{get_genes_matching_regex_motif}
\title{Get yeast genes matching a regular expression}
\usage{

  get_genes_matching_regex_motif(motif_regexp = "c(g|a|v|l|i){2,2}(m|s|q|a|c|l|e)\\\\*",
  genome_http_dir = "http://downloads.yeastgenome.org/sequence/S288C_reference/orf_protein/orf_trans_all.fasta.gz",
  genome_dest_dir = "../data/yeast_seq.gz")
}
\arguments{
\item{motif_regexp}{Desired regular expression}

\item{genome_http_dir}{Web address of genome, a gzipped multi-fasta file}

\item{genome_dest_dir}{Where you want genome downloaded}
}
\value{
names of yeast genes (ORF name - e.g. YKR103W) matching desired expression
}
\description{
Downloads yeast genome to a given directory
if it does not already exist, and checks the genome
(a gzipped multi-fasta file) for genes matching a given
regular expression
}

