% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{get_largest_component_size}
\alias{get_largest_component_size}
\title{Given a PCA file, returns the size of the largest connected component for a given condition}
\usage{
get_largest_component_size(pca_file, condition)
}
\arguments{
\item{pca_file}{a list of interactions, split by Condition in the Condition columm
and each pair indexed by ORF.1 and ORF.2}

\item{condition}{in what condition the largest connected component is needed (e.g. FK506)}
}
\value{
the size of the largest connected component (numeric)
}
\description{
Given a PCA file, returns the size of the largest connected component for a given condition
}

