% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_analysis_utilities.R
\name{given_motif_enrichment}
\alias{given_motif_enrichment}
\title{Test for conditional motif enrichment}
\usage{
given_motif_enrichment(universe_file, pca_call_file, motif_regexp, condition)
}
\arguments{
\item{universe_file}{A data frame which contains all the pairs
in the PCA experiment under the columns 'ORF.1','ORF.2'}

\item{pca_call_file}{A data frame which contains all signficant calls
in the PCA experiment (either enhanced or depleted), 
with the condition of interest under the
'Condition' column and the orf pairs under the columns 'ORF.1','ORF.2'}

\item{motif_regexp}{The regular expression to search for}

\item{condition}{the name of the condition, as it appears in pca_call_file}
}
\description{
Calculates possible enrichment for a given motif
under a given condition
}

