% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{keep_only_largest_connected_component}
\alias{keep_only_largest_connected_component}
\title{Return largest connected component}
\usage{
keep_only_largest_connected_component(graph)
}
\arguments{
\item{graph}{and igraph graph object}
}
\value{
the subgraph consisting of the largest connected component in the graph
}
\description{
Return largest connected component
}

