% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{make_network_iterations}
\alias{make_network_iterations}
\title{Simulate a network that is the same size as one obtained from an experimental file}
\usage{
make_network_iterations(pca_universe, pca_enhanced_file, pca_depleted_file,
  condition, n_iters, mode = "edgewise", node_sensitivity = 1,
  prob_node = 1, metric = get_largest_component_from_edgelist,
  n_parallel = 4, cluster = NULL)
}
\arguments{
\item{pca_universe}{A PCA file corresponding to all protein-protein interactions
available in a given condition}

\item{condition}{the condition of interest, a string present in both pca_universe and pca_file}

\item{n_iters}{number of simulations to make}

\item{mode}{if 'edgewise', samples edges directly (much faster), if 'nodewise'
samples a proportion of nodes and edges given by prob_node}

\item{node_sensitivity}{in nodewise sampling, what is the probability that a given PPI will be
captured by the assay? (numeric between 0 and 1)}

\item{prob_node}{if mode is 'nodewise', what proportion of the sampled interactions
should be from a sampled node compared to a sampled edge}

\item{metric}{summary statistic for the returned network - should work on an edgelist
(a two colum matrix), if not specified, returns a list of two-column matrices}

\item{n_parallel}{number of parallel computing sessions to start}

\item{cluster}{the identity of a cluster created by the SNOW package, if not specified, it creates ones}

\item{pca_file}{A PCA file corresponding to all significantly changed
protein-protein interactions}
}
\value{
a list of two-column matrices if metric is not specified, or a list of the output of metric
on the two column-matrix if it is
}
\description{
Simulate a network that is the same size as one obtained from an experimental file
}

