% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_action_prediction.R
\name{mass_action_predictor}
\alias{mass_action_predictor}
\title{A simple mass action prediction for protein complex level changes}
\usage{
mass_action_predictor(c1, c2, fc1, fc2, k = NA, concentration_factor = 20,
  mode = "analytic")
}
\arguments{
\item{c1}{concentration of the first protein in the PPI pair, in arbitrary units}

\item{c2}{concentration of the second protein in the PPI pair, in arbitrary units}

\item{k}{affinity estimate (in same units as c1,c2), set to min(c1,c2)/concentration factor by default}

\item{concentration_factor}{used for unknown Kd assignment, defaults to 20}

\item{mode}{either analytic for an exact solution (default) or iterative}

\item{r1}{ratio of concentration change of the first protein}

\item{r2}{ratio of concentration change of the second protein}
}
\value{
a vector representing the predicted complex level ratio
}
\description{
A simple mass action prediction for protein complex level changes
}

