% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{network_connectivity_graph}
\alias{network_connectivity_graph}
\title{Plots a graph corresponding to the largest connected components in the enhanced and depleted
PPIs for a given PCA object}
\usage{
network_connectivity_graph(pca_universe, pca_enhanced, pca_depleted, condition,
  color_scale, default_node_color = rgb(0.3, 0.3, 0.3), node_size = 10,
  edge_width = 7, min_fc = -1, max_fc = 1, node_text_colour = "white",
  seed = 104, edit = F, load_saved = T, to_plot = "both",
  layout_save_dir = "", layout_file = "default.layout",
  network_export_file = "default.tsv", my_title = "\\n\\nDoxorubicin")
}
\arguments{
\item{pca_universe}{A PCA file showing all possible interactions for any given condition}

\item{pca_enhanced}{A PCA file showing enhanced interactions for each condition}

\item{pca_depleted}{A PCA file showing depleted interactions for each condition}

\item{condition}{the condition the scale should be plotted for}

\item{color_scale}{a list of colours to be made into a gradient}

\item{default_node_color}{what to colour the nodes if they are not to be made into a colour scale}

\item{node_size}{}

\item{edge_width}{}

\item{min_fc}{minimum value for colour scale}

\item{max_fc}{maximum "}

\item{node_text_colour}{tex value for nodes}

\item{seed}{random seed}

\item{edit}{Should graph be edited prior to plotting? Opens a tkplot window to allow editing of plot}

\item{load_saved}{Should a prior layout be loaded?}

\item{layout_save_dir}{Location of prior layour (absolute path)}

\item{layout_file}{Name of layout file}

\item{my_title}{what to title the plot?}
}
\value{
A plot of the largest connected component in a given condition
}
\description{
Plots a graph corresponding to the largest connected components in the enhanced and depleted
PPIs for a given PCA object
}

