% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{network_simulation_significance}
\alias{network_simulation_significance}
\title{Returns a matrix corresponding to significant deviations from an observed network, given a defined property}
\usage{
network_simulation_significance(pca_universe, pca_enhanced, pca_depleted,
  excluded_condition_grep = "CRISPR", iterations = 1000, seed = 123,
  sigval = 0.05, mode = "edgewise", node_sensitivity = 1, prob_node = 1,
  metric = "component_size", n_parallel = 8, cluster = NULL)
}
\arguments{
\item{pca_universe}{A PCA file corresponding to all protein-protein interactions
available in a given condition}

\item{pca_enhanced}{A PCA file corresponding to all significantly enhanced
protein-protein interactions}

\item{pca_depleted}{A PCA file corresponding to all significantly depleted
protein-protein interactions}

\item{excluded_condition_grep}{a regular expression which matches all conditions to be excluded from the matrix}

\item{iterations}{number of sampled iterations to simulate}

\item{seed}{seed for random number generator}

\item{sigval}{significance cutoff (default 0.05)}

\item{mode}{either 'edgewise' or 'nodewise', whether the simulations should just sample edges from the main
network or sample a combination of edges, and nodes by which to obtain edges}

\item{node_sensitivity}{if a node is sampled, what proportion of the edges are recovered on average?}

\item{prob_node}{what proportion of edges in a given simulation should be obtained by sampling PPIs form nodes}

\item{metric}{what summary statistic of the real and simulated network should be used to evaluate significance?}

\item{n_parallel}{number of parallel processes to spawn for simulation}

\item{cluster}{a pre-defined SNOW cluster, default is to create a new one}
}
\value{
a matrix where the rows correspond to directions and the columns are conditions,
values are either 1 (significant, real value is greater than simulations), 0 (nonsigificant), -1
(significant, real value is less than simulations)
}
\description{
Returns a matrix corresponding to significant deviations from an observed network, given a defined property
}

