% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{network_simulation_significance_node_edge_search_matrix}
\alias{network_simulation_significance_node_edge_search_matrix}
\title{Performs a search over different proportions of node-based (indirect) protein complex changes in
a subnetwork simulation, and returns a matrix in which the rows are all condition x direction combinations,
and the columns correspond to the given proportions of node-based (indirect) protein complex changes}
\usage{
network_simulation_significance_node_edge_search_matrix(pca_universe,
  pca_enhanced, pca_depleted, iterations = 100, node_probs = c(0:10)/10,
  metric = "component size", conditions = NULL,
  excluded_condition_grep = "CRISPR", n_parallel = 8, load_saved = F,
  save_output = F, save_directory = NULL, save_filename = NULL,
  seed = 99)
}
\arguments{
\item{pca_universe}{A PCA file corresponding to all protein-protein interactions
available in a given condition}

\item{pca_enhanced}{A PCA file corresponding to all significantly enhanced
protein-protein interactions}

\item{pca_depleted}{A PCA file corresponding to all significantly depleted
protein-protein interactions}

\item{iterations}{number of iterations}

\item{node_probs}{a vector indicating the proportions of node-based interaction changes to be simulated}

\item{metric}{the metric to be tested for significance, e.g. largest component size or graph density}

\item{conditions}{conditions to be tested, defaults to all in pca_universe}

\item{excluded_condition_grep}{a regular expression which matches all conditions to be excluded from the matrix}

\item{n_parallel}{number of parallel processes to spawn for the simulations}

\item{load_saved}{if True, attempts to load a previous run instead of executing from the path specified by save_directory and save_filename}

\item{save_output}{if True, will save the output to the path specified by save_directory and save_filename}

\item{save_directory}{the directory of where the files will be loaded/saved}

\item{save_filename}{the filename of where the files will be loaded/saved}

\item{seed}{seed for the random number generator}
}
\value{
a matrix where the rows correspond to conditions x directions and the columns are given proportions of node-based changes,
values are either 1 (significant, real value is greater than simulations), 0 (nonsigificant), -1
(significant, real value is less than simulations)
}
\description{
Performs a search over different proportions of node-based (indirect) protein complex changes in
a subnetwork simulation, and returns a matrix in which the rows are all condition x direction combinations,
and the columns correspond to the given proportions of node-based (indirect) protein complex changes
}

