% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{node_edge_search_heatmap}
\alias{node_edge_search_heatmap}
\title{Returns a heatmap which plots the output from network_simulation_significance_node_edge_search_matrix}
\usage{
node_edge_search_heatmap(node_edge_sig_matrix, my_color_function,
  legend_labels = c(`-1` = "Decreased\\ncomponent size", `0` =
  "Expected\\ncomponent size", `1` = "Increased\\ncomponent size"),
  x_label = "",
  y_label = "Proportion of Protein-\\nCentered Complex Changes",
  border_colour = "grey40", zero_colour = "grey90", border_size = 0.5)
}
\arguments{
\item{node_edge_sig_matrix}{the output from network_simulation_significance_node_edge_search_matrix,
where the rows correspond to conditions x directions and the columns are given proportions of node-based changes,
values are either 1 (significant, real value is greater than simulations), 0 (nonsigificant), -1
(significant, real value is less than simulations)}

\item{my_color_function}{a gradient generating function, the first value will be used for -1 values in matrix,
the middle for 0 values, the last for 1 values}

\item{legend_labels}{vector of form c('-1'=label1,'0'=label2,'1'=label3)}

\item{x_label}{the title to be plotted on the X axis}

\item{y_label}{the title to be plotted on the Y axis}

\item{border_colour}{border outline colour}

\item{border_size}{border outline width}
}
\value{
plots a heatmap
}
\description{
Returns a heatmap which plots the output from network_simulation_significance_node_edge_search_matrix
}

