% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_action_prediction.R
\name{pca_ma_precision_plot}
\alias{pca_ma_precision_plot}
\title{Plots precisions for mass-action based complex predictions as a function of cutoff}
\usage{
pca_ma_precision_plot(my_predictions, output_path, filename, draw = F,
  prediction_cutoffs = "AUTO", p_cutoff = 0.05, effect_size_cutoff = 0.25,
  bottom_predition_limit = -2.5, top_predition_limit = 2.5,
  line_width = 3, bootstrap_iters = 1000, enhanced_colour = rgb(0.75, 0.1,
  0.1), depleted_colour = rgb(0.1, 0.1, 0.75), polygon_transparency = 0.3,
  label_size = 1.3, legend_x_position = -1.2, legend_y_position = 0.85,
  xlabel = "mRNA Predicted Log2(R) Cutoff", ylabel = "Precision",
  legend_labels = c("Accumulated complexes", "Depleted complexes"))
}
\arguments{
\item{my_predictions}{output table from pca_ma_prediction}

\item{output_path}{where the plot will be saved if draw is False}

\item{filename}{the filename of the plot if draw is False}

\item{draw}{if True, plots the file, if False, saves a CairoPDF object}

\item{prediction_cutoffs}{the resolution of the plot, defaults to full}

\item{p_cutoff}{the adjusted p value cutoff for calling significant interactions}

\item{effect_size_cutoff}{the adjusted effect size cutoff for caling significant interactions}

\item{bottom_predition_limit}{lowest effect size plotted}

\item{top_predition_limit}{highest effect size plotted}

\item{bootstrap_iters}{how many bootstraps to draw the 95% polygon}
}
\description{
Plots precisions for mass-action based complex predictions as a function of cutoff
}

