% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_action_prediction.R
\name{pca_ma_prediction}
\alias{pca_ma_prediction}
\title{Provides mass-action based predictions for bcPCA output in a specific format}
\usage{
pca_ma_prediction(pca_file, abundance_file, expression_file, condition,
  expression_control_regexp = "Ethanol.0h",
  expression_condition_regexp = "Ethanol.4h")
}
\arguments{
\item{pca_file}{filename corresponding to output from the PCA experiment,
unfiltered, with both UPtag and DNtag measurements available}

\item{abundance_file}{filename corresponding to the abundance of each ORF in the PCA
file, first column of file should be the ORF ID, the second column should be a measure of
abundance (in arbitrary units)}

\item{expression_file}{mRNA expression data from which the predictions are going to be made,
must have an ORF column corresponding to the ORF, and the experiments must be organized in columns}

\item{condition}{the condition in the pca_file which is to be correlated}

\item{expression_control_regexp}{regular expression used to match the control condition in the mRNA file}

\item{expression_condition_regexp}{regular expression used to match the condition of interesti n the mRNA file}
}
\value{
a data frame with the original data combined with the predictions
}
\description{
Provides mass-action based predictions for bcPCA output in a specific format
}

