% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_action_prediction.R
\name{pca_ma_prediction_plot}
\alias{pca_ma_prediction_plot}
\title{Plots mRNA-based bcPCA predictions compared to quantitative bc-PCA measurements}
\usage{
pca_ma_prediction_plot(my_predictions, output_path, draw = F,
  filename = "bcPCA_mRNA_predictions.pdf",
  prediction_colname = "Log2_MA_prediction",
  measurement_colname = "bcPCA_FC.AVG", point_colours = rgb(0.08, 0.08,
  0.35, 0.3), outline_colours = rgb(0, 0, 0, 0.1), point_size = 0.5,
  label_size = 1.3, xlimits = c(-5, 5), ylimits = c(-4, 2),
  relative_text_position_x = 0.2, relative_text_position_y = 0.8,
  x_axis_label = "mRNA Predicted Log2(R)",
  y_axis_label = "Measured Log2(R)", long_tick_length = 0.04,
  short_tick_length = 0.02)
}
\arguments{
\item{my_predictions}{prediction data frame, output from pca_ma_prediction}

\item{output_path}{where the plot is to be written}

\item{filename}{the file name of the plot'}
}
\description{
Plots mRNA-based bcPCA predictions compared to quantitative bc-PCA measurements
}

