% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_action_prediction.R
\name{pca_mRNA_error_analysis}
\alias{pca_mRNA_error_analysis}
\title{Performs an error analysis of the mRNA based BCA predictions, based on the reproducibility
of UP and DN tag measurements in the PCA experiment and (by default)
the reproducibility of the 4h and 12h mRNA based measurements}
\usage{
pca_mRNA_error_analysis(pca_file, expression_file, abundance_file,
  condition = "ethanol", q_val_cutoff = 1, fc_cutoff = 0,
  up_dn_diff_cutoff = Inf, expression_condition_regexp_vec = c("Ethanol.4h",
  "Ethanol.1h"), expression_control_regexp_vec = c("Ethanol.0h",
  "Ethanol.0h"), iters = 100, my_seed = 1234)
}
\arguments{
\item{pca_file}{the location of a PCA file}

\item{expression_file}{the location of an expression file}

\item{abundance_file}{the location of a file encoding protein abundance}

\item{condition}{condition tested, so far expression only works on ethanol}

\item{q_val_cutoff}{q value cutoff for PCA experiment}

\item{fc_cutoff}{log2(fold change) cutoff for PCA}

\item{expression_condition_regexp_vec}{regular expression vector that matches the conditions in the expression experiment to be compared}

\item{expression_control_regexp_vec}{regular expression vector that matches the reference conditions in the expression experiment by which the conditions are to be compared to}
}
\value{
a matrix of noisy mRNA-based PCA measurements
}
\description{
Performs an error analysis of the mRNA based BCA predictions, based on the reproducibility
of UP and DN tag measurements in the PCA experiment and (by default)
the reproducibility of the 4h and 12h mRNA based measurements
}

