% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_connectivity_permutation.R
\name{set_network_aesthethics}
\alias{set_network_aesthethics}
\title{Sets graph aesthetics for an igraph object - nodes are given a singe colour, edges are given multiple}
\usage{
set_network_aesthethics(graph, node_size, node_colour, node_text_colour,
  edge_width, edge_attribute, edge_colour_scale, edge_colour_min,
  edge_colour_max, edge_ncolours = 100, text_size_constant = 25)
}
\arguments{
\item{graph}{igraph object}

\item{node_size}{size to plot nodes}

\item{node_colour}{colour for all nodes}

\item{node_text_colour}{text colour for all nodes}

\item{edge_width}{width to plot edges}

\item{edge_attribute}{attribute by which to set the edge colour scale by}

\item{edge_colour_scale}{the colour scale object for the edge - a list of colours e.g. c('red','green')}

\item{edge_colour_min}{the minimum value for the colour scale}

\item{edge_colour_max}{the maximum value for the colour scale}

\item{edge_ncolours}{how many colours in the gradient}

\item{text_size_constant}{used to set character size based on number of characters, higher = smaller text}
}
\value{
an igraph graph with the given graphics parameters modified
}
\description{
Sets graph aesthetics for an igraph object - nodes are given a singe colour, edges are given multiple
}

