% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment_utilities.R
\name{vector_to_python_list}
\alias{vector_to_python_list}
\title{Converts a vector to a literal python list representation (all items in strings)}
\usage{
vector_to_python_list(contents)
}
\arguments{
\item{contents}{a vector of string}
}
\value{
a list which contains one item per line in the file name
}
\description{
Converts a vector to a literal python list representation (all items in strings)
}
\examples{
vector_to_python_list(c('A','B'))
"['A','B']"
}

